+++ components/src/core/collapsible-header-layout/collapsible-header-layout.tsx
@@ -10,7 +10,8 @@ import {
 } from 'react-native';
 import { useTheme } from 'react-native-paper';
 import { $DeepPartial } from '@callstack/react-theme-provider';
-import { ANIMATION_LENGTH, Header, HeaderProps as PXBHeaderProps, heightWithStatusBar } from '../header';
+import { ANIMATION_LENGTH, Header, HeaderProps as PXBHeaderProps } from '../header';
+import { useHeaderDimensions } from '../hooks/useHeaderDimensions';
 
 export type CollapsibleLayoutProps = ViewProps & {
     /** Props to spread to the Header component. */
@@ -49,12 +50,15 @@ export const CollapsibleHeaderLayout: React.FC<CollapsibleLayoutProps> = (props)
 
     const theme = useTheme(themeOverride);
     const scrollRef = useRef(null);
+
+    const { getScaledHeight } = useHeaderDimensions();
+
     const animatedScrollValue = useRef(new Animated.Value(0)).current;
     const [scrollValue, setScrollValue] = useState(0);
     const headerVariant = HeaderProps.variant || 'dynamic';
     const startExpanded = HeaderProps.startExpanded || false;
-    const collapsedHeight = heightWithStatusBar(HeaderProps.collapsedHeight || 56);
-    const expandedHeight = heightWithStatusBar(HeaderProps.expandedHeight || 200);
+    const collapsedHeight = getScaledHeight(HeaderProps.collapsedHeight || 56);
+    const expandedHeight = getScaledHeight(HeaderProps.expandedHeight || 200);
     const scrollableDistance = expandedHeight - collapsedHeight;
     const initialScrollPosition = headerVariant === 'static' ? 0 : !startExpanded ? scrollableDistance : 0;