+++ components/src/core/drawer/drawer-nav-item.tsx
@@ -11,6 +11,7 @@ import { useNavGroupContext } from './context/nav-group-context';
 import { findChildByType, inheritSharedProps } from './utilities';
 import * as Colors from '@pxblue/colors';
 import Collapsible from 'react-native-collapsible';
+import { EdgeInsets, useSafeAreaInsets } from 'react-native-safe-area-context';
 
 export type DrawerNavItemStyles = {
     root?: StyleProp<ViewStyle>;
@@ -103,7 +104,8 @@ export type NavItem = DrawerNavItemProps;
 export type NestedNavItem = NestedDrawerNavItemProps;
 
 // First nested item has no additional indentation.  All items start with 16px indentation.
-const calcNestedPadding = (depth: number): number => (depth > 0 ? (depth - 1) * 32 : 0);
+const calcNestedPadding = (depth: number, insets: EdgeInsets): number =>
+    insets.left + (depth > 0 ? (depth - 1) * 32 : 0);
 
 const makeStyles = (
     props: DrawerNavItemProps,
@@ -235,6 +237,8 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
         // other View props
     } = otherProps;
 
+    const insets = useSafeAreaInsets();
+
     const [expanded, setExpanded] = useState(isInActiveTree); // isInActiveTree: there is a bug in the react-native-collapsible that incorrectly calculates the initial panel height when using nested collapse panels
     const active = activeItem === itemID;
     const hasAction = Boolean(onItemSelect || onPress || (items && items.length > 0) || Boolean(children));
@@ -301,7 +305,7 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
     );
 
     const infoListItemStyles = styles.infoListItem || {};
-    const { root: iliRoot, title: iliTitle, ...otherILI } = infoListItemStyles;
+    const { root: iliRoot, title: iliTitle, statusStripe: iliStatusStripe, ...otherILI } = infoListItemStyles;
 
     return (
         <>
@@ -330,7 +334,7 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
                             onPress={hasAction ? onPressAction : undefined}
                             hidePadding={hidePadding}
                             styles={{
-                                root: Object.assign({ paddingLeft: calcNestedPadding(depth) }, iliRoot),
+                                root: Object.assign({ paddingLeft: calcNestedPadding(depth, insets) }, iliRoot),
                                 title: Object.assign(
                                     active || (isInActiveTree && !disableActiveItemParentStyles)
                                         ? {
@@ -343,6 +347,7 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
                                           },
                                     iliTitle
                                 ),
+                                statusStripe: Object.assign({ left: insets.left }, iliStatusStripe),
                                 ...otherILI,
                             }}
                         />